Isolate durable conversation contracts
The durable conversation module mixed API schemas with append and recovery behavior, making the remaining decomposition work harder to review. Move schema and type contracts into a dedicated module while keeping durable.ts as the compatibility export surface.

Constraint: Existing imports from src/agent/conversation/durable.ts must keep working.

Rejected: Split append recovery in the same PR | too broad for one roadmap slice.

Confidence: high

Scope-risk: narrow

Directive: Keep durable-contracts.ts behavior-free; append, resync, and queue behavior belong in narrower runtime modules.

Tested: deno test --no-check --allow-all src/agent/conversation/durable-contracts.test.ts src/agent/conversation/durable.test.ts src/agent/conversation/run-chunk-mirror.test.ts src/agent/conversation/run-mirror.test.ts src/agent/conversation/run-stream-mirror.test.ts

Tested: deno check src/agent/conversation/durable-contracts.ts src/agent/conversation/durable-contracts.test.ts src/agent/conversation/durable.ts src/agent/conversation/durable.test.ts src/agent/conversation/hosted-lifecycle.ts src/agent/conversation/hosted-terminal.ts src/agent/conversation/run-chunk-mirror.ts src/agent/conversation/run-mirror.ts src/agent/conversation/run-stream-mirror.ts

Tested: git diff --check
